+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Deprecated global validateEmail() in favor of widget.validate.email()
 * Move (global) jquery plugins (sortlist, destroyed event) from widget-util to widget-core
 * EXPERIMENTAL - form  and input field validation
+* EXPERIMENTAL - styles
 
 
 0.7.2 (2016.03.10)